refactor (NormalPitches) : 기존 메소드 삭제  리팩토링
