+++ src/test/java/bowling/domain/bowling/BowlingPinsGroupTest.java
@@ -5,21 +5,14 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class BowlingPinsGroupTest {
 
     @AfterEach
     public void resetBowlingPins() {
         BowlingPinsGroup.initiate().hitByBall(10);
-    }
-
-    @DisplayName("매 프레임의 첫 번째 투구 등 처음 칠 때 볼링 핀을 초기화 시켜 반환함")
-    @Test
-    public void initiateBowlingPinsGroup_초기화() {
-        assertThatCode(() -> {
-            BowlingPinsGroup.initiate();
-        }).doesNotThrowAnyException();
     }
 
     @DisplayName("다음 투구로 넘어가는 next 요청을 보내면, 현재 넘어지지 않은 볼링핀만을 모아 리턴함")
