+++ src/test/java/bowling/domain/frame/FinalFrameTest.java
@@ -1,3 +1,4 @@
+/*
 package bowling.domain;
 
 import bowling.domain.frame.FinalFrame;
@@ -27,12 +28,14 @@ public class FinalFrameTest {
         assertThat(finalFrame.next()).isEqualTo(null);
     }
 
-    /*
+    */
+/*
     FinalPitch의 종료 조건
     1) 2번 투구시 미스인 경우 -> 종료
     2) 스트라이크 인 경우 -> 2번 더 칠 수 있음. (그 2번이 스트라이크여도됨.)
     3) 스페어인 경우 -> 1번 더 칠 수 있음. (그 1번은 스트라이크여도됨.)
-     */
+     *//*
+
 
     @DisplayName("스페어의 경우 1번 더 칠 수 있도록 현재 Frame 반환")
     @Test
@@ -54,3 +57,4 @@ public class FinalFrameTest {
         assertThat(finalFrame.next()).isEqualTo(null);
     }
 }
+*/
